Struct crater::mmheap::MmHeap

pub struct MmHeap<T> { /* private fields */ }

An implicit binary heap that can efficiently find the min or max of its contained elements.

Find min / find max: O(1)

Pop min / pop max: O(log(n))

Push: O(log(n))

Heapify: O(n) The comparison function can be changed after creation by just calling MmHeap::ify_by again.

Implementations

impl<T> MmHeap<T>

pub fn new() -> Self

Create an empty MmHeap

pub fn make(buf: Vec<T>, cmp: &impl Fn(&T, &T) -> Ordering) -> Self

Create an MmHeap out of a vector and immediately heapify it according to cmp Note that cmp takes references, which means the common case of storing references in the heap can require wrapping the comparison function since it will have to take references to references

pub fn ify_by(&mut self, cmp: &impl Fn(&T, &T) -> Ordering)

Reorder the heap according to a new comparison function. It’s impossible to add to the heap’s private buffer directly, so this is only necessary when changing the comparison function. When calling MmHeap::push etc with the same comparison function, the heap will maintain its invariant

pub fn len(&self) -> usize

Get the number of elements in the heap

pub fn peek_min(&self) -> Option<&T>

Get the minimum element without removing it

pub fn peek_max_by(&self, cmp: &impl Fn(&T, &T) -> Ordering) -> Option<&T>

Get the maximum element without removing it

pub fn push_by(&mut self, e: T, cmp: &impl Fn(&T, &T) -> Ordering)

Insert an element into the heap Elements that compare equal are fine, but their order will be unspecified

pub fn pop_min_by(&mut self, cmp: &impl Fn(&T, &T) -> Ordering) -> Option<T>

Get the minimal element and remove it

pub fn pop_max_by(&mut self, cmp: &impl Fn(&T, &T) -> Ordering) -> Option<T>

Get the maximal element and remove it

pub fn pushpop_min_by(&mut self, e: T, cmp: &impl Fn(&T, &T) -> Ordering) -> T

Insert a new element and remove the min element in the resulting heap in a single operation Could be faster than MmHeap::push_by + MmHeap::pop_min_by separately (the current implementation is only more efficient if the heap is empty, but it will always be at least as good)

pub fn pushpop_max_by(&mut self, e: T, cmp: &impl Fn(&T, &T) -> Ordering) -> T

Insert a new element and remove the max element in the resulting heap in a single operation

pub fn extend_by<U: IntoIterator<Item = T>>( &mut self, iter: U, cmp: &impl Fn(&T, &T) -> Ordering )